[refactor] mdDiary, mdTrash 패키지 구조 변경
- 기능별로 분리
