+++ src/main/java/com/fcc/PureSync/context/diary/dto/MdDiaryRequestDto.java
@@ -1,21 +1,17 @@
-package com.fcc.PureSync.dto;
+package com.fcc.PureSync.context.diary.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class MdDiaryResponseDto {
-    Long dySeq;
+public class MdDiaryRequestDto {
     String dyDate;
     String dyTitle;
     String dyContents;
     String emoState;
-    int emoField;
 }
