+++ src/main/java/com/fcc/PureSync/context/diary/dto/MdDiaryResponseDto.java
@@ -1,17 +1,21 @@
-package com.fcc.PureSync.dto;
+package com.fcc.PureSync.context.diary.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
+
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class MdDiaryRequestDto {
+public class MdDiaryResponseDto {
+    Long dySeq;
     String dyDate;
     String dyTitle;
     String dyContents;
     String emoState;
+    int emoField;
 }
