+++ src/main/java/com/fcc/PureSync/context/diary/entity/Emotion.java
@@ -1,4 +1,4 @@
-package com.fcc.PureSync.entity;
+package com.fcc.PureSync.context.diary.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
